+++ src/views/loading.blade.php
@@ -1,0 +1,7 @@
+@extends('cruddy::layout')
+
+@section('content')
+    <span class="loading">
+        @lang('cruddy::app.loading')
+    </span>
+@stop